+++ lib/aipp/scopes/notam/executable.rb
@@ -6,7 +6,7 @@ module AIPP
       def options
         AIPP.options.merge(
           module: 'NOTAM',
-          effective_at: now - now.sec - (now.min * 60)   # previous full hour
+          effective_at: Time.now.change(min: 0, sec: 0)
         )
       end
 
@@ -15,14 +15,12 @@ module AIPP
           Download online NOTAM and convert it to #{AIPP.options.schema.upcase}.
           Usage: #{File.basename($0)} notam [options]
         END
-        o.on('-t', '--effective (TIME)', String, %Q[effective after this point in time (default: #{AIPP.options.effective_at})]) { AIPP.options.effective_at = Time.parse(_1) }
+        o.on('-t', '--effective (TIME)', String, %Q[effective at this time (default: "#{AIPP.options.effective_at}")]) { AIPP.options.effective_at = Time.parse(_1) }
         o.on('-x', '--crossload DIR', String, 'crossload directory') { AIPP.options.crossload = Pathname(_1) }
       end
 
-      private
-
-      def now
-        @now ||= Time.now.utc.round
+      def guard
+        AIPP.options.effective_at = AIPP.options.effective_at.change(min: 0, sec: 0).utc
       end
 
     end