+++ lib/aipp/scopes/shoot/parser.rb
@@ -1,0 +1,9 @@
+module AIPP
+  module SHOOT
+
+    # @abstract
+    class Parser < AIPP::Parser
+    end
+
+  end
+end